+++ crates/db_schema/src/aggregates/mod.rs
@@ -1,5 +1,3 @@
 #[cfg(feature = "full")]
 pub mod person_post_aggregates;
-#[cfg(feature = "full")]
-pub mod site_aggregates;
 pub mod structs;
